Criando entidades, criando DTO`s

+++ KeepHealth.API/KeepHealth.Domain/Doctor.cs
@@ -5,7 +5,7 @@ namespace KeepHealth.Domain
     public class Doctor : BaseEntity
     {
         public List<MedicalSpeciality> MedicalSpecialities { get; set; }
-        public decimal AppointmentPrice { get; set; }
+        public double AppointmentPrice { get; set; }
         public User User { get; set; }
         public bool Approved { get; set; }
     }
